--- la-gloria-store/src/HomeComponents/ProductCard.js
@@ -1,20 +1,0 @@
-import React from 'react';
-
-const ProductCard = () => {
-    return (
-        <div className="card" style={{ width: '18rem' }}>
-            <img src="https://2.bp.blogspot.com/-_Tx3nbVr5mM/U7ctz8a3LDI/AAAAAAAAAkM/YDDEPDIHjVE/s1600/escudoo.png" className="card-img-top" alt="..." />
-            <div className="card-body">
-                <h5 className="card-title">Card title</h5>
-                <p className="card-text">
-                    Some quick example text to build on the card title and make up the bulk of the card's content.
-                </p>
-                <a href="#" className="btn btn-primary">
-                    Go somewhere
-                </a>
-            </div>
-        </div>
-    );
-};
-
-export default ProductCard;
