--- la-gloria-store/src/HomeComponents/ProductGrid.js
@@ -1,23 +1,0 @@
-import React from 'react';
-import ProductCard from './ProductCard';
-
-function ProductGrid() {
-    const gridStyle = {
-        display: 'grid',
-        gridTemplateColumns: 'repeat(3, 1fr)',
-        gap: '20px',
-    };
-
-    return (
-        <div style={gridStyle}>
-            <ProductCard />
-            <ProductCard />
-            <ProductCard />
-            <ProductCard />
-            <ProductCard />
-            <ProductCard />
-        </div>
-    );
-}
-
-export default ProductGrid;
