+++ la-gloria-store/src/pages/Home.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ProductTable from '../HomeComponents/ProductTable';
+import ProductTable from '../HomeComponents/ProductTable/ProductTable';
 import '../App.css';
 import Carousel from "../HomeComponents/Carousel";
